fix(storage): store ids in a leaf module, break the boot import cycle
clientStorage imports the boot payload, which imports the platform registry, which read the store ids back from clientStorage; evaluated in that order the registry hit a temporal dead zone and the mock build booted to a blank page. the ids now live in storeIds.ts, re-exported by clientStorage, and the registry imports them from there.
